fix scheduler queue name, oof
turns out the scheduled queue is named "scheduler" and not "schedule",
and it is very important to make sure that it actually runs. this change
delegates directly to `config/sidekiq.yml` for the first/default sidekiq
process, and then copies the queue list and priorities from that file
(without the scheduler queue) for the sidekiq process that can be scaled
up.